fixed a date type bug

+++ bond_energy.cpp
@@ -1,6 +1,6 @@
 #include "header.hpp"
 
-double bond_energy (std::array<Eigen::MatrixXi, 2> &spin_lattice, int s, int i, int j, std::map<std::string, double> & params) {
+double bond_energy (std::array<Eigen::MatrixXd, 2> &spin_lattice, int s, int i, int j, std::map<std::string, double> & params) {
     // for this lattice, we have two sub lattice, label by s, which blongs to {0, 1}
     // and i, j denote the site of sub lattice s.
     // so the spin of (i, j) site of s sub lattice is spin_lattice[s](i, j)
