+++ header.hpp
@@ -15,6 +15,6 @@
 
 void read_para(std::map<std::string, double> &params);
 // void init_conf(Eigen::MatrixXi &spin_conf, const int Nx, const int Ny, );
-void init_conf(Eigen::MatrixXi &spin_conf, const int Nx, const int Ny, boost::random::mt19937 & seed_spin);
-double bond_energy(std::array<Eigen::MatrixXi, 2> &spin_lattice, int s, int i, int j, std::map<std::string, double> & params);
-void sweep(std::array<Eigen::MatrixXi, 2> & spin_lattice, const int Nx, const int Ny, std::map<std::string, double> & params,  boost::random::mt19937 & rng);
+void init_conf(Eigen::MatrixXd &spin_conf, const int Nx, const int Ny, boost::random::mt19937 & seed_spin);
+double bond_energy(std::array<Eigen::MatrixXd, 2> &spin_lattice, int s, int i, int j, std::map<std::string, double> & params);
+void sweep(std::array<Eigen::MatrixXd, 2> & spin_lattice, const int Nx, const int Ny, std::map<std::string, double> & params,  boost::random::mt19937 & rng);
